requirements.txt
blinker==1.6.2
boto3==1.28.22
botocore==1.31.22
bson==0.5.10
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
Click==7.0
cryptography==41.0.3
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-SocketIO==4.2.1
google-auth==1.27.0
google-auth-oauthlib==0.4.2
greenlet==0.4.15
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==1.0.1
MarkupSafe==1.1.1
marshmallow==3.19.0
monotonic==1.5
oauthlib==3.1.0
packaging==23.1
pathlib==1.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.4.1
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-engineio==3.11.1
python-socketio==4.4.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.9
s3transfer==0.6.1
six==1.13.0
stripe==5.5.0
typing_extensions==4.7.1
urllib3==1.26.16
vendor==0.1
Werkzeug==0.16.0